Extract DITA utility methods to DitaUtils
