+++ src/main/java/org/dita/dost/reader/MergeTopicParser.java
@@ -8,17 +8,6 @@
  */
 package org.dita.dost.reader;
 
-import static javax.xml.XMLConstants.*;
-import static org.dita.dost.chunk.ChunkModule.isLocalScope;
-import static org.dita.dost.reader.MergeMapParser.*;
-import static org.dita.dost.util.Constants.*;
-import static org.dita.dost.util.FileUtils.*;
-import static org.dita.dost.util.URLUtils.*;
-import static org.dita.dost.util.URLUtils.setFragment;
-
-import java.io.File;
-import java.net.URI;
-
 import org.dita.dost.log.DITAOTLogger;
 import org.dita.dost.util.Job;
 import org.dita.dost.util.MergeUtils;
@@ -28,6 +17,17 @@ import org.xml.sax.Attributes;
 import org.xml.sax.SAXException;
 import org.xml.sax.helpers.AttributesImpl;
 import org.xml.sax.helpers.XMLFilterImpl;
+
+import java.io.File;
+import java.net.URI;
+
+import static javax.xml.XMLConstants.XML_NS_URI;
+import static org.dita.dost.reader.MergeMapParser.ATTRIBUTE_NAME_OHREF;
+import static org.dita.dost.reader.MergeMapParser.ATTRIBUTE_NAME_OID;
+import static org.dita.dost.util.Constants.*;
+import static org.dita.dost.util.DitaUtils.isLocalScope;
+import static org.dita.dost.util.FileUtils.stripFragment;
+import static org.dita.dost.util.URLUtils.*;
 
 /**
  * MergeTopicParser reads topic file and transform the references to other dita
