+++ src/main/java/org/dita/dost/writer/DitaWriterFilter.java
@@ -21,8 +21,8 @@ import java.io.File;
 import java.net.URI;
 import java.util.Map;
 
-import static org.dita.dost.chunk.ChunkModule.isLocalScope;
 import static org.dita.dost.util.Constants.*;
+import static org.dita.dost.util.DitaUtils.isLocalScope;
 import static org.dita.dost.util.URLUtils.*;
 import static org.dita.dost.reader.GenListModuleReader.*;
