+++ src/main/java/org/dita/dost/writer/ForceUniqueFilter.java
@@ -20,8 +20,8 @@ import java.util.Deque;
 import java.util.HashMap;
 import java.util.Map;
 
-import static org.dita.dost.chunk.ChunkModule.isLocalScope;
 import static org.dita.dost.util.Constants.*;
+import static org.dita.dost.util.DitaUtils.isLocalScope;
 import static org.dita.dost.util.FileUtils.getExtension;
 import static org.dita.dost.util.FileUtils.replaceExtension;
 import static org.dita.dost.util.URLUtils.*;
